Fix issue with creating a profile not waiting for login success.
When creating an account, a user is to be redirected to their profile;
however, this requires the user to be logged in. This was being performed,
but the result of logging in was not received before issuing a request to
get profile information first. This has been corrected to await the login
response before navigating to the user profile.